Refactor the code to prepare for concurrency
This rearranges the code a bit.  Dispatch is moved from request.rs to
session.rs, while code the reads and writes from the channel (which was
distributed between request.rs and session.rs before) now all resides in
channel.rs.  The channel hidden behind a "port" and a "chan" named
for (but not implemented with) Rust's standard Port and Chan.